fix: encode URLs entered in the text field
Previously, query parameters/fragments in these URLs would be interpreted as part of the url of the page and not the site being documented.

That would then break URLs like `https://cdn.skypack.dev/react?dts`, because the `?dts` would be dropped and cause the raw JS to be documented with no types.